Strip the manual pinnings from all dependencies
There no longer is a need for having manual pinnings of our
dependencies as the global pinning file holds all of these. So remove
them from the recipe. Re-rendering will pull in the appropriate pinnings
for into local variant files, which will be used for each build and
determine the constraints of the resultant packages.